src/requirements.txt
python-binance
pandas
openpyxl
websocket
plotly